matrix-animations-web/backend/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==4.4.2
Flask==2.2.3
Flask-Cors==3.0.10
fonttools==4.39.3
gunicorn==20.1.0
idna==3.4
imageio==2.28.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.5.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
moviepy==1.0.3
numpy==1.24.2
packaging==23.1
Pillow==9.5.0
proglog==0.1.10
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.29.0
six==1.16.0
tqdm==4.65.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
